requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coverage==5.1
cryptography==2.9
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-graphql-jwt==0.3.1
djangorestframework==3.11.0
drf-jwt==1.15.2
Faker==0.9.1
graphene==2.1.8
graphene-django==2.9.1
graphql-core==3.1.0
graphql-relay==3.0.0
gunicorn==20.0.4
idna==2.9
mixer==6.1.3
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
promise==2.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
Rx==3.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.1.9
whitenoise==5.0.1
